encapsule validate and info edittext
