+++ organize/app/src/main/res/values/strings.xml
@@ -64,7 +64,7 @@
     <string name="message_info_password">Entre 6 a 10 caracteres alfanuméricos.</string>
     <string name="message_info_cpf">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Phasellus eget sem quis risus sagittis tincidunt nec tempus libero. Suspendisse non.</string>
     <string name="message_info_birth_date">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Phasellus eget sem quis risus sagittis tincidunt nec tempus libero. Suspendisse non.</string>
-    <string name="message_validate_code_plan">Código de confirmação enviado em seu e-mail na contratação do plano. Caso não tenha recebido o codigo ou não o solicitou, entre em contato com o suporte Organize ou acesse com o plano gratuito.</string>
+    <string name="message_info_code_plan">Código de confirmação enviado em seu e-mail na contratação do plano. Caso não tenha recebido o codigo ou não o solicitou, entre em contato com o suporte Organize ou acesse com o plano gratuito.</string>
     <string name="message_confirmation_forgot_password">E-mail de recuperação de senha enviado com sucesso.</string>
 
     <string name="placeholder_cpf">Insira o seu CPF.</string>
